requirements.txt
apache-airflow==2.10.5
snowflake-sqlalchemy==1.7.3
snowflake-connector-python==3.13.2
psycopg2-binary==2.9.5